update contract type

+++ backend/src/main/java/com/logumrh/repository/common/ContractTypeRepository.java
@@ -3,6 +3,6 @@ package com.logumrh.repository.common;
 import com.logumrh.model.common.ContractType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContractTypeRepository extends JpaRepository<ContractType, Integer> {
+public interface ContractTypeRepository extends JpaRepository<ContractType, Long> {
 
 }
